requirements.txt
asgiref==3.3.1
Django==3.1.7
django-cart==1.0.4
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
install==1.3.4
Pillow
psycopg2-binary
PyJWT==1.7.1
pytz==2021.1
sqlparse==0.4.1
